requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-rest-swagger==0.3.5
django-toolbelt==0.0.1
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
drf-extensions==0.2.8
gunicorn==19.4.5
psycopg2==2.6.1
PyJWT==1.4.0
PyYAML==3.11
static3==0.6.1
